+++ DfE.FindInformationAcademiesTrusts/Pages/Trusts/Academies/OfstedRatings.cshtml.cs
@@ -1,5 +1,6 @@
 using DfE.FindInformationAcademiesTrusts.Data;
 using DfE.FindInformationAcademiesTrusts.Data.Enums;
+using DfE.FindInformationAcademiesTrusts.Services;
 using DfE.FindInformationAcademiesTrusts.Services.Academy;
 using DfE.FindInformationAcademiesTrusts.Services.DataSource;
 using DfE.FindInformationAcademiesTrusts.Services.Trust;
@@ -7,16 +8,17 @@ using Microsoft.AspNetCore.Mvc;
 
 namespace DfE.FindInformationAcademiesTrusts.Pages.Trusts.Academies;
 
-public class OfstedRatingsModel : TrustsAreaModel, IAcademiesAreaModel
+public class OfstedRatingsModel : AcademiesPageModel
 {
     public AcademyOfstedServiceModel[] Academies { get; set; } = default!;
     private IAcademyService AcademyService { get; }
 
     public OfstedRatingsModel(ITrustProvider trustProvider, IDataSourceService dataSourceService,
-        ILogger<OfstedRatingsModel> logger, ITrustService trustService, IAcademyService academyService) : base(
-        trustProvider, dataSourceService, trustService, logger, "Academies in this trust")
+        ILogger<OfstedRatingsModel> logger, ITrustService trustService, IAcademyService academyService, IExportService exportService, IDateTimeProvider dateTimeProvider) : base(
+        trustProvider, dataSourceService, trustService, exportService, logger, dateTimeProvider)
     {
         PageTitle = "Academies Ofsted ratings";
+        TabName = "Ofsted ratings";
         AcademyService = academyService;
     }
 
@@ -40,6 +42,4 @@ public class OfstedRatingsModel : TrustsAreaModel, IAcademiesAreaModel
 
         return pageResult;
     }
-
-    public string TabName => "Ofsted ratings";
 }