+++ DfE.FindInformationAcademiesTrusts/Pages/Trusts/Academies/PupilNumbers.cshtml.cs
@@ -1,5 +1,6 @@
 using DfE.FindInformationAcademiesTrusts.Data;
 using DfE.FindInformationAcademiesTrusts.Data.Enums;
+using DfE.FindInformationAcademiesTrusts.Services;
 using DfE.FindInformationAcademiesTrusts.Services.Academy;
 using DfE.FindInformationAcademiesTrusts.Services.DataSource;
 using DfE.FindInformationAcademiesTrusts.Services.Trust;
@@ -7,18 +8,18 @@ using Microsoft.AspNetCore.Mvc;
 
 namespace DfE.FindInformationAcademiesTrusts.Pages.Trusts.Academies;
 
-public class PupilNumbersModel : TrustsAreaModel, IAcademiesAreaModel
+public class PupilNumbersModel : AcademiesPageModel
 {
     public IAcademyService AcademyService { get; }
     public AcademyPupilNumbersServiceModel[] Academies { get; set; } = default!;
 
     public PupilNumbersModel(ITrustProvider trustProvider, IDataSourceService dataSourceService,
-        ILogger<PupilNumbersModel> logger, ITrustService trustService, IAcademyService academyService) : base(
-        trustProvider, dataSourceService,
-        trustService, logger, "Academies in this trust")
+        ILogger<PupilNumbersModel> logger, ITrustService trustService, IAcademyService academyService, IExportService exportService, IDateTimeProvider dateTimeProvider)
+        : base(trustProvider, dataSourceService, trustService, exportService, logger, dateTimeProvider)
     {
         AcademyService = academyService;
         PageTitle = "Academies pupil numbers";
+        TabName = "Pupil numbers";
     }
 
     public override async Task<IActionResult> OnGetAsync()
@@ -34,8 +35,6 @@ public class PupilNumbersModel : TrustsAreaModel, IAcademiesAreaModel
 
         return pageResult;
     }
-
-    public string TabName => "Pupil numbers";
 
     public static string PhaseAndAgeRangeSortValue(AcademyPupilNumbersServiceModel academy)
     {