+++ tests/DfE.FindInformationAcademiesTrusts.UnitTests/Pages/Trusts/Academies/AcademiesDetailsModelTests.cs
@@ -2,6 +2,7 @@ using DfE.FindInformationAcademiesTrusts.Data;
 using DfE.FindInformationAcademiesTrusts.Data.Enums;
 using DfE.FindInformationAcademiesTrusts.Pages;
 using DfE.FindInformationAcademiesTrusts.Pages.Trusts.Academies;
+using DfE.FindInformationAcademiesTrusts.Services;
 using DfE.FindInformationAcademiesTrusts.Services.Academy;
 using DfE.FindInformationAcademiesTrusts.Services.Trust;
 using DfE.FindInformationAcademiesTrusts.UnitTests.Mocks;
@@ -16,6 +17,8 @@ public class AcademiesDetailsModelTests
     private readonly Mock<ITrustProvider> _mockTrustProvider = new();
     private readonly Mock<ITrustService> _mockTrustRepository = new();
     private readonly Mock<IAcademyService> _mockAcademyService = new();
+    private readonly Mock<IExportService> _mockExportService = new();
+    private readonly Mock<IDateTimeProvider> _mockDateTimeProvider = new();
     private readonly MockDataSourceService _mockDataSourceService = new();
     private readonly MockLogger<AcademiesDetailsModel> _mockLogger = new();
 
@@ -27,8 +30,8 @@ public class AcademiesDetailsModelTests
             .ReturnsAsync(_fakeTrust);
 
         _sut = new AcademiesDetailsModel(_mockTrustProvider.Object, _mockDataSourceService.Object,
-                _mockLinkBuilder.Object, _mockLogger.Object, _mockTrustRepository.Object, _mockAcademyService.Object)
-            { Uid = _fakeTrust.Uid };
+                _mockLinkBuilder.Object, _mockLogger.Object, _mockTrustRepository.Object, _mockAcademyService.Object, _mockExportService.Object, _mockDateTimeProvider.Object)
+        { Uid = _fakeTrust.Uid };
     }
 
     [Fact]