+++ tests/DfE.FindInformationAcademiesTrusts.UnitTests/Pages/Trusts/Academies/FreeSchoolMealsModelTests.cs
@@ -2,6 +2,7 @@ using DfE.FindInformationAcademiesTrusts.Data;
 using DfE.FindInformationAcademiesTrusts.Data.Enums;
 using DfE.FindInformationAcademiesTrusts.Data.UnitTests.Mocks;
 using DfE.FindInformationAcademiesTrusts.Pages.Trusts.Academies;
+using DfE.FindInformationAcademiesTrusts.Services;
 using DfE.FindInformationAcademiesTrusts.Services.Trust;
 using DfE.FindInformationAcademiesTrusts.UnitTests.Mocks;
 using Microsoft.AspNetCore.Mvc;
@@ -14,6 +15,8 @@ public class FreeSchoolMealsModelTests
     private readonly Mock<IFreeSchoolMealsAverageProvider> _mockFreeSchoolMealsAverageProvider = new();
     private readonly Mock<ITrustProvider> _mockTrustProvider = new();
     private readonly Mock<ITrustService> _mockTrustRepository = new();
+    private readonly Mock<IExportService> _mockExportService = new();
+    private readonly Mock<IDateTimeProvider> _mockDateTimeProvider = new();
     private readonly MockDataSourceService _mockDataSourceService = new();
 
     public FreeSchoolMealsModelTests()
@@ -27,8 +30,8 @@ public class FreeSchoolMealsModelTests
 
         _sut = new FreeSchoolMealsModel(_mockTrustProvider.Object, _mockFreeSchoolMealsAverageProvider.Object,
                 _mockDataSourceService.Object, new MockLogger<FreeSchoolMealsModel>().Object,
-                _mockTrustRepository.Object)
-            { Uid = "1234" };
+                _mockTrustRepository.Object, _mockExportService.Object, _mockDateTimeProvider.Object)
+        { Uid = "1234" };
     }
 
     [Fact]