+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased][unreleased]
 
+
+### Added
+
+- Academies in this trust data now exportable
+
 ### Changed
 
 - Reduced the load time of the Academies in trust pupil numbers page